app/requirements.txt
fastapi>=0.109.1
uvicorn>=0.22.0
qbittorrent-api>=2023.10.54
Jinja2
python-dotenv
